Remove `user_is_not_requester` validation from Purchase steps
* Was causing Micro-purchase requests to not have steps because Alan D
  is both the MP requester and Purchaser
* Closes https://rpm.newrelic.com/accounts/921394/applications/5480870/traced_errors/2713dca3-5e58-11e6-9456-f8bc12425498_0_18136?original_error_id=2713dca3-5e58-11e6-9456-f8bc12425498_0_18136
